# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NexaShift AI Earnings Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /daily-plan/{userId}:
    get:
      operationId: getDailyPlan
      tags: [decisions]
      summary: Get AI-generated daily work plan
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
...
  /ai/explain:
    get:
      operationId: getAiExplanation
      tags: [ai]
      summary: Get AI explanation for a prediction
      parameters:
        - name: event
          in: query
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: AI explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExplanation"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [insights]
      summary: Get aggregated dashboard summary for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
components:
  schemas:
...
    DailyPlan:
      type: object
      properties:
        userId:
          type: string
        date:
          type: string
        bestWorkingHours:
          type: array
          items:
            $ref: "#/components/schemas/WorkWindow"
        highDemandZones:
          type: array
          items:
            $ref: "#/components/schemas/Zone"
        riskWindows:
          type: array
          items:
            $ref: "#/components/schemas/RiskWindow"
        expectedEarnings:
          type: number
        recommendation:
          type: string
        weatherForecast:
          type: string
        topTip:
          type: string
      required:
        - userId
        - date
        - bestWorkingHours
        - highDemandZones
        - riskWindows
        - expectedEarnings
        - recommendation
        - weatherForecast
        - topTip
    WorkWindow:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        demandScore:
          type: number
        estimatedEarning:
          type: number
        label:
          type: string
      required:
        - startTime
        - endTime
        - demandScore
        - estimatedEarning
        - label
    Zone:
      type: object
      properties:
        name:
          type: string
        demandScore:
          type: number
        avgEarning:
          type: number
        distanceKm:
          type: number
      required:
        - name
        - demandScore
        - avgEarning
        - distanceKm
    RiskWindow:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        riskType:
          type: string
        severity:
          type: string
        advice:
          type: string
      required:
        - startTime
        - endTime
        - riskType
        - severity
        - advice
...
    AiExplanation:
      type: object
      properties:
        event:
          type: string
        factors:
          type: array
          items:
            $ref: "#/components/schemas/ExplanationFactor"
        summary:
          type: string
        confidence:
          type: number
        historicalAccuracy:
          type: number
      required:
        - event
        - factors
        - summary
        - confidence
        - historicalAccuracy
    ExplanationFactor:
      type: object
      properties:
        factor:
          type: string
        impact:
          type: string
        weight:
          type: number
        description:
          type: string
      required:
        - factor
        - impact
        - weight
        - description
    DashboardSummary:
      type: object
      properties:
        userId:
          type: string
        todayEarnings:
          type: number
        weeklyEarnings:
          type: number
        monthlyEarnings:
          type: number
        currentRiskScore:
          type: number
        activePolicies:
          type: integer
        totalClaimsThisMonth:
          type: integer
        totalPayoutThisMonth:
          type: number
        riskTrend:
          type: string
          enum: [improving, stable, worsening]
        incomeProtected:
          type: number
        streakDays:
          type: integer
      required:
        - userId
        - todayEarnings
        - weeklyEarnings
        - monthlyEarnings
        - currentRiskScore
        - activePolicies
        - totalClaimsThisMonth
        - totalPayoutThisMonth
        - riskTrend
        - incomeProtected
        - streakDays